src/requirements.txt
discord.py>=2.0.1
asyncpg>=0.27.0
pytz>=2022.5
python-dotenv>=0.21.0
